Move button for creating child notes to the note header
